Finished creating services for the creation of the dictionary.json file to be used in the app
