+++ src/js/i18n/lang-list.json
@@ -1,0 +1,10 @@
+{
+  "eng": {
+    "name": "English", 
+    "file": "src/js/i18n/lang/english.json"
+  },
+  "esp": {
+    "name": "Espa\u00F1ol",
+    "file": "src/js/i18n/lang/spanish.json"
+  }
+}
