+++ src/js/i18n/lang/english.json
@@ -1,0 +1,125 @@
+{ 
+  "STRING-101": "Welcome!",
+  "STRING-102": "Get started by creating a new wallet,<br> open or import your existing wallet.",
+  "STRING-103": "Open wallet",
+  "STRING-104": "Open your existing wallet file (ctrl+o)",
+  "STRING-105": "Create New Wallet",
+  "STRING-106": "Create new wallet file (ctrl+n)",
+  "STRING-107": "Import key",
+  "STRING-108": "Import/restore your private keys (ctrl+i)",
+  "STRING-109": "<p>This is a GUI front-end to manage ICX assets (ICX).</p><p>You can use this software to create and import new ICX wallets, or open existing wallets. Then you can manage your wallet, such as viewing your balance and transactions history, sending/transferring ICX, etc.</p>",
+  "STRING-110": "Donate to: <strong>ICONation</strong>",
+  "STRING-111": "Click to copy",
+  "STRING-112": "Source codes, release notes, documentations & bugs tracker can be found at our Github repo:<br> <a tabindex='0' class='external' href='https://github.com/ICONation/svalinn'>https://github.com/ICONation/svalinn</a>",
+  "STRING-113": "Add Address Book Entry",
+  "STRING-114": "Name",
+  "STRING-115": "Required, give a unique name/label to make it easy to identify...",
+  "STRING-116": "Wallet Address",
+  "STRING-117": "Required, a valid ICX address, non unique address is allow as long as it have different payment id...",
+  "STRING-118": "Save",
+  "STRING-119": "Back",
+  "STRING-120": "Address Book",
+  "STRING-121": "Search",
+  "STRING-122": "Search by name or wallet address",
+  "STRING-123": "Select address book",
+  "STRING-124": "Default/Built-in Address Book",
+  "STRING-125": "Create new address book",
+  "STRING-126": "Add New Entry",
+  "STRING-127": "STRING-119",
+  "STRING-128": "Create ICX transaction",
+  "STRING-129": "Wallet password",
+  "STRING-130": "Password to open this wallet...",
+  "STRING-131": "Recipient address",
+  "STRING-132": "Required, a valid ICX address...",
+  "STRING-133": "Type one or more characters to search from address book, or type new address.",
+  "STRING-134": "Amount (ICX)",
+  "STRING-135": "Required, amount of ICX you want to send.",
+      "STRING-136": "Step Limit (step)",
+      "STRING-137": "Required, step price is paid in loop and 1 loop is fixed to 0.000000000000000001 (10^18) ICX. ICON transaction fee is imposed according to various factors such as the number of smart contract usage, the amount of blockchain database used and the size of transaction data, etc",
+      "STRING-138": "Select where the transaction file will be stored",
+      "STRING-139": "Required, full path to store this transaction...",
+      "STRING-140": "Select network",
+      "STRING-141": "Mainnet",
+      "STRING-142": "Testnet for Exchanges (Euljiro)",
+      "STRING-143": "Testnet for DApps (Yeouido)",
+      "STRING-144": "Create",
+      "STRING-145": "STRING-119",
+      "STRING-146": "Svalinn - ICONation ICX Wallet",
+      "STRING-147": "Wallet Overview",
+      "STRING-148": "Wallet",
+      "STRING-149": "STRING-120",
+      "STRING-150": "STRING-120",
+      "STRING-151": "Create transaction",
+      "STRING-152": "STRING-151",
+      "STRING-153": "Send transaction",
+      "STRING-154": "STRING-153",
+      "STRING-155": "Language",
+      "STRING-156": "About",
+      "STRING-157": "Show keyboard shorcut",
+      "STRING-158": "Switch to dark mode",
+      "STRING-159": "STRING-105",
+      "STRING-160": "Select where the wallet file will be stored",
+      "STRING-161": "Required, full path to store this wallet...",
+      "STRING-162": "Set password to open this wallet",
+      "STRING-163": "Required, set a password to open your new wallet...",
+      "STRING-164": "STRING-144",
+      "STRING-165": "Cancel",
+      "STRING-166": "Address:",
+      "STRING-167": "STRING-111",
+      "STRING-168": "Balance:",
+      "STRING-169": "Available Balance",
+      "STRING-170": "STRING-140",
+      "STRING-171": "STRING-141",
+      "STRING-172": "STRING-142",
+      "STRING-173": "STRING-143",
+      "STRING-174": "Backup Keys",
+      "STRING-175": "Backup your private keys (ctrl+e)",
+      "STRING-176": "Close",
+      "STRING-177": "Close current wallet, so you can open another wallet (ctrl+x)",
+      "STRING-178": "Import Private Key",
+      "STRING-179": "STRING-160",
+      "STRING-180": "STRING-161",
+      "STRING-181": "STRING-162",
+      "STRING-182": "Required",
+      "STRING-183": "STRING-163",
+      "STRING-184": "Private key to be imported",
+      "STRING-185": "Required, your private view key to be imported...",
+      "STRING-186": "Import",
+      "STRING-187": "STRING-165",
+      "STRING-188": "Open a wallet",
+      "STRING-189": "Select the wallet file",
+      "STRING-190": "Full path to the wallet file to be opened...",
+      "STRING-191": "Open",
+      "STRING-192": "STRING-119",
+      "STRING-193": "Reveal or export your Private Key",
+      "STRING-194": "Important!!!<br>Please be sure to backup your keys and store them somewhere private and safe.<br>If your wallet file is corrupted or lost you can always import these keys to restore the wallet in any compatible program.",
+      "STRING-195": "STRING-129",
+      "STRING-196": "STRING-130",
+      "STRING-197": "Your private key:",
+      "STRING-198": "Click the Reveal button to see it",
+      "STRING-199": "Reveal",
+      "STRING-200": "Export to file",
+      "STRING-201": "STRING-119",
+      "STRING-202": "Send ICX transaction",
+      "STRING-203": "Select the transaction file",
+      "STRING-204": "Full path to the transaction file to be opened...",
+      "STRING-205": "Transaction Hash:",
+      "STRING-206": "Click to see the transaction on the tracker",
+      "STRING-207": "Send",
+      "STRING-208": "STRING-119",
+      "STRING-209": "Available Keybindings", 
+      "STRING-210": "Switch to <strong>overview/welcome</strong> screen", 
+      "STRING-211": "Switch to <strong>next screen</strong>", 
+      "STRING-212": "STRING-211", 
+      "STRING-213": "Switch to <strong>Create new wallet</strong> screen", 
+      "STRING-214": "Switch to <strong>Open a wallet</strong> screen", 
+      "STRING-215": "Switch to <strong>Import wallet from private keys</strong> screen", 
+      "STRING-216": "Switch to <strong>Export private keys</strong> screen (when wallet opened)", 
+      "STRING-217": "Switch to <strong>Create Transaction</strong> screen (when wallet opened)", 
+      "STRING-218": "Switch to <strong>Send Transactions</strong> screen (when wallet opened)", 
+      "STRING-219": "Close wallet", 
+      "STRING-220": "Toggle dark/night mode", 
+      "STRING-221": "Display shortcut key information (this dialog)", 
+      "STRING-222": "Close any opened dialog (like this dialog)"
+}
+
