+++ src/js/i18n/lang/spanish.json
@@ -1,0 +1,125 @@
+{ 
+  "STRING-101": "Bienvenido!",
+  "STRING-102": "Inicia creando una nueva billetera,<br> Abre o importa tu billetera.",
+  "STRING-103": "Abrir billetera",
+  "STRING-104": "Abrir tu archivo de billetera existente (ctrl+o)",
+  "STRING-105": "Crear nueva billetera",
+  "STRING-106": "Crear nuevo archivo de billetera (ctrl+n)",
+  "STRING-107": "Importar llave",
+  "STRING-108": "Importar/restaurar tus llaves privadas (ctrl+i)",
+  "STRING-109": "<p>Este es un programa para el manejo de activos de ICX (ICX).</p><p>Usted puede usar este programa para crear e importar billeteras de ICX o para abrir billeteras ya existentes. Luego puede gestionar sus billeteras, observando el balance, historial de transacciones, enviar/transferir ICX, etc.</p>",
+  "STRING-110": "Donar a: <strong>ICONation</strong>",
+  "STRING-111": "Click para copiar",
+  "STRING-112": "Puede conseguir codigos fuente, notas de lanzamiento, documentaciones y rastreador de bugs en nuestro repositorio de Github:<br> <a tabindex='0' class='external' href='https://github.com/ICONation/svalinn'>https://github.com/ICONation/svalinn</a>",
+  "STRING-113": "Adicionar direccion al directorio",
+  "STRING-114": "Nombre",
+  "STRING-115": "Requerido, adicionar nombre o etiqueta unica para identificar facilmente...",
+  "STRING-116": "Direccion de Billetera",
+  "STRING-117": "Requerido, direccion de ICX valida, direcciones no unicas son permitidas siempre y cuando tengan diferentes ID de pago...",
+  "STRING-118": "Guardar",
+  "STRING-119": "Atras",
+  "STRING-120": "Directorio",
+  "STRING-121": "Buscar",
+  "STRING-122": "Buscar por nombre o direccion de billetera",
+  "STRING-123": "Seleccionar directorio",
+  "STRING-124": "Directorio predeterminado",
+  "STRING-125": "Crear nuevo directorio",
+  "STRING-126": "Agregar nueva entrada",
+  "STRING-127": "STRING-119",
+  "STRING-128": "Crear transaccion de ICX",
+  "STRING-129": "Clave de billetera",
+  "STRING-130": "Clave para abrir esta billetera...",
+  "STRING-131": "Direccion de recipiente",
+  "STRING-132": "Requerido, direccion valida de ICX...",
+  "STRING-133": "Tipee uno o mas caracteres para buscar en el directorio, o agregue una nueva direccion.",
+  "STRING-134": "Monto (ICX)",
+  "STRING-135": "Requerido, cantidad de ICX que desea enviar.",
+      "STRING-136": "Limite de Step (step)",
+      "STRING-137": "Requerido, el precio del 'step' es pagado en loop (bucle), y 1 'loop' esta fijado a 0.000000000000000001 (10^18) ICX. EL costo de la transaccion de ICON esta impuesto de acuerdo a varios factores tales como, el numero de contratos inteligentes en uso, la cantidad de base de datos del blockchain usados y el tama\u00F1o de los datos de la transaccion, etc.",
+      "STRING-138": "Selecciona donde sera guardado el archivo de la transaccion",
+      "STRING-139": "Requerido, ruta completa para guardar esta transaccion...",
+      "STRING-140": "Seleccionar red",
+      "STRING-141": "Mainnet",
+      "STRING-142": "Testnet para Exchanges (Euljiro)",
+      "STRING-143": "Testnet para DApps (Yeouido)",
+      "STRING-144": "Crear",
+      "STRING-145": "STRING-119",
+      "STRING-146": "Svalinn - Billetera de ICX de ICONation",
+      "STRING-147": "Resumen de billetera",
+      "STRING-148": "Billetera",
+      "STRING-149": "STRING-120",
+      "STRING-150": "STRING-120",
+      "STRING-151": "Crear transaccion",
+      "STRING-152": "STRING-151",
+      "STRING-153": "Enviar transaccion",
+      "STRING-154": "STRING-153",
+      "STRING-155": "Lenguage",
+      "STRING-156": "Acerca de",
+      "STRING-157": "Mostrar atajos del teclado",
+      "STRING-158": "Cambiar a modo nocturno",
+      "STRING-159": "STRING-105",
+      "STRING-160": "Seleccionar donde guardar el archivo de la billetera",
+      "STRING-161": "Requerido, ruta completa donde guardar esta billetera...",
+      "STRING-162": "Configurar clave para abrir esta billetera",
+      "STRING-163": "Requerido, configurar una clave para abrir tu nueva billetera...",
+      "STRING-164": "STRING-144",
+      "STRING-165": "Cancelar",
+      "STRING-166": "Direccion:",
+      "STRING-167": "STRING-111",
+      "STRING-168": "Balance:",
+      "STRING-169": "Balance disponible",
+      "STRING-170": "STRING-140",
+      "STRING-171": "STRING-141",
+      "STRING-172": "STRING-142",
+      "STRING-173": "STRING-143",
+      "STRING-174": "Respaldar llaves",
+      "STRING-175": "Respaldar tus llaves privadas (ctrl+e)",
+      "STRING-176": "Cerrar",
+      "STRING-177": "Cerrar esta billetera para poder abrir otra (ctrl+x)",
+      "STRING-178": "Importar llave privada",
+      "STRING-179": "STRING-160",
+      "STRING-180": "STRING-161",
+      "STRING-181": "STRING-162",
+      "STRING-182": "Requerido",
+      "STRING-183": "STRING-163",
+      "STRING-184": "Llave privada que desea importar",
+      "STRING-185": "Requerido, llave privada a importar...", 
+      "STRING-186": "Importar",
+      "STRING-187": "STRING-165",
+      "STRING-188": "Abrir billetera",
+      "STRING-189": "Seleccionar archivo de billetera",
+      "STRING-190": "Ruta completa del archivo de la billetera a abrir...",
+      "STRING-191": "Abrir",
+      "STRING-192": "STRING-119",
+      "STRING-193": "Mostrar o importar su llave privada",
+      "STRING-194": "Importante!!!<br>Por favor asegurese de respaldar sus llaves y guardarlas en un lugar privado y seguro.<br>Si el archivo de su billetera esta corrompido o perdido usted siempre puede importar estas llaves para restaurar la billetera en cualquier programa compatible.",
+      "STRING-195": "STRING-129",
+      "STRING-196": "STRING-130",
+      "STRING-197": "Su llave privada:",
+      "STRING-198": "Click en 'Mostrar' para ver",
+      "STRING-199": "Mostrar",
+      "STRING-200": "Exportar a un archivo",
+      "STRING-201": "STRING-119",
+      "STRING-202": "Enviar transaccion de ICX",
+      "STRING-203": "Seleccionar el archivo de transaccion",
+      "STRING-204": "Ruta completa del archivo de transaccion a abrir...",
+      "STRING-205": "Hash de transaccion:",
+      "STRING-206": "Click para ver la transaccion en el tracker",
+      "STRING-207": "Enviar",
+      "STRING-208": "STRING-119",
+      "STRING-209": "Atajos de teclado disponibles", 
+      "STRING-210": "Cambiar a pantalla de <strong>Resumen/Bienvenida</strong>", 
+      "STRING-211": "Cambiar a <strong>Siguiente Pantalla</strong>", 
+      "STRING-212": "STRING-211", 
+      "STRING-213": "Cambiar a pantalla <strong>Crear nueva Billetera</strong>", 
+      "STRING-214": "Cambiar a pantalla <strong>Abrir Billetera</strong>", 
+      "STRING-215": "Cambiar a pantalla <strong>Importar billetera desde llave privada</strong>", 
+      "STRING-216": "Cambiar a pantalla <strong>Exportar llaves privadas</strong> (cuando billetera esta abierta)", 
+      "STRING-217": "Cambiar a pantalla <strong>Crear transaccion</strong> (cuando billetera esta abierta)", 
+      "STRING-218": "Cambiar a pantalla <strong>Enviar transaccion</strong> (cuando billetera esta abierta)", 
+      "STRING-219": "Cerrar billetera", 
+      "STRING-220": "Activar/Desactivar modo nocturno", 
+      "STRING-221": "Mostrar informacion de atajo de teclado (este dialogo)", 
+      "STRING-222": "Cerrar cualquier dialogo abierto (como este dialogo)"
+}
+
